Integrate with travis-ci.org
We had to reorganize tests in order to trigger green travis ci jobs.
This is because some tests required the dev database. "integration"
tests are now called "front". Now test_synchro without any ugly hack.

This proves that unit tests can now run without any dev database, which
is a huge win \o/
