+++ labonneboite/importer/conf/test.py
@@ -1,5 +1,7 @@
+import os
+
 # --- importer input directory of DPAE and ETABLISSEMENT exports
-INPUT_SOURCE_FOLDER = '/srv/lbb/labonneboite/importer/tests/data'
+INPUT_SOURCE_FOLDER = os.path.join(os.path.dirname(__file__), '..', '..', 'tests', 'importer', 'data')
 
 # --- job 1/8 & 2/8 : check_etablissements & extract_etablissements
 DISTINCT_DEPARTEMENTS_HAVING_OFFICES = 15
