+++ labonneboite/tests/web/front/test_admin.py
