+++ labonneboite/tests/web/front/test_delete_user_account.py
