+++ labonneboite/tests/web/front/test_favorites.py
