+++ labonneboite/tests/web/front/test_static_pages.py
