+++ vagrant/playbook.yml
@@ -17,6 +17,7 @@
     - name: Install APT packages.
       apt: name={{ item }} state=present
       with_items:
+        - language-pack-fr
         - git
         - python
         - python-dev
